ai_girlfriend/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0
aiohttp==3.9.1
python-multipart==0.0.6
chromadb==0.4.18
pandas==2.1.4
numpy==1.24.3
pillow==10.2.0
pydub==0.25.1
schedule==1.2.0
python-dateutil==2.8.2
tzlocal==5.2
colorlog==6.8.0
tqdm==4.66.1